requirements.txt
rembg
sumy
nltk
textblob
pillow
opencv-python
moviepy
